Implemented Dynamic Routes to render every product detail page seperately and without hardcoding it
We can think about this as the pages opens up when you click to a product in a page. The general idea of the page rendered and informations might be taken from a database and implemented into a different component like ProductDetail.js with more information parts and design etc.